the big cleanup
Fortunately, the ids are not really colliding, so I can just move on and this will only be noticeable to very few people. It's annoying that I lose my progress locally, but OK, maybe a reason to retest and redo the path myself, haha. It should not feel like a drag, right? Forcing myself through it again is a good reality check. Goodbye you AI generated stuff.